backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Evan Mitchell",email = "emitchell@developertown.com"}
]
readme = "README.md"
requires-python = ">=3.9, <4.0"
dependencies = [
    "requests (>=2.32.4,<3.0.0)",
    "openai (>=1.86.0,<2.0.0)",
    "onnxruntime (==1.16.3)",
    "chromadb (==1.0.11)",
    "protobuf (==3.20.3)",
    "flask (>=3.1.1,<4.0.0)",
    "flask-restx (>=1.3.0,<2.0.0)",
    "flask-smorest (>=0.46.1,<0.47.0)",
    "flask-cors (>=6.0.1,<7.0.0)",
    "umap-learn (>=0.5.9.post2,<0.6.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
